requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
pinecone==5.3.1
boto3==1.35.50
python-docx==1.1.2
PyMuPDF==1.24.12
docx2txt==0.8
python-dotenv==1.0.1
